refactor(app): update directory structure
