--- src/about/components/about.component.ts
@@ -1,8 +1,0 @@
-import {Component} from 'angular2/core';
-
-@Component({
-  selector: 'sd-about',
-  templateUrl: './about/components/about.component.html',
-  styleUrls: ['./about/components/about.component.css']
-})
-export class AboutComponent {}
