+++ src/client/app/+about/index.ts
@@ -1,0 +1,2 @@
+export * from './components/about.component';
+
