+++ src/client/app/+home/index.ts
@@ -1,0 +1,2 @@
+export * from './components/home.component';
+
