+++ src/client/app/components/app.component.ts
@@ -2,9 +2,9 @@ import {Component} from 'angular2/core';
 import {ROUTER_DIRECTIVES, RouteConfig} from 'angular2/router';
 import {NavbarComponent} from './navbar.component';
 import {ToolbarComponent} from './toolbar.component';
-import {HomeComponent} from '../../home/components/home.component';
-import {AboutComponent} from '../../about/components/about.component';
-import {NameListService} from '../../shared/services/name-list.service';
+import {NameListService} from '../shared/index';
+import {HomeComponent} from '../+home/index';
+import {AboutComponent} from '../+about/index';
 
 @Component({
   selector: 'sd-app',
@@ -13,7 +13,15 @@ import {NameListService} from '../../shared/services/name-list.service';
   directives: [ROUTER_DIRECTIVES, NavbarComponent, ToolbarComponent]
 })
 @RouteConfig([
-  { path: '/',      name: 'Home',  component: HomeComponent  },
-  { path: '/about', name: 'About', component: AboutComponent }
+  {
+    path: '/',
+    name: 'Home',
+    component: HomeComponent
+  },
+  {
+    path: '/about',
+    name: 'About',
+    component: AboutComponent
+  }
 ])
 export class AppComponent {}
