+++ src/client/app/shared/index.ts
@@ -1,0 +1,2 @@
+export * from './services/name-list.service';
+
