+++ src/client/app/shared/services/name-list.service.ts
