+++ src/client/main.ts
