+++ tools/config/project.config.ts
@@ -16,11 +16,5 @@ export class ProjectConfig extends SeedConfig {
     const seedDependencies = this.NPM_DEPENDENCIES;
 
     this.NPM_DEPENDENCIES = seedDependencies.concat(additional_deps);
-
-    this.APP_ASSETS = [
-      // {src: `${this.ASSETS_SRC}/css/toastr.min.css`, inject: true},
-      // {src: `${this.APP_DEST}/assets/scss/global.css`, inject: true},
-      { src: `${this.ASSETS_SRC}/main.css`, inject: true },
-    ];
   }
 }
